refactor(engine-host): extract the CORE engine cluster below the seam
The final, atomic step of #120: move the mutually-recursive runtime/session core
out of rusty-sudocode-cli/src/main.rs into engine_host::{runtime_build,
session_engine}, so the CLI is purely a renderer and no longer names the
engine-internal turn types.

runtime_build: RuntimeConfig, RuntimePluginState, BuiltRuntime (+Deref/DerefMut/
Drop, wrapping ConversationRuntime<EngineApiClient, CliToolExecutor>), and the
build_runtime* chain (plugins, MCP, permission policy, system prompt,
EngineApiClient). session_engine: SessionEngine (inherent impl + impl
EngineDelegate [whose run_turn was the last thing naming runtime::RuntimeObserver]
+ impl SessionLifecycle), the SessionLifecycle trait, AcpCliSession,
ModelSwitchReport.

Boundary preserved, no cycle:
- engine-host gains telemetry + commands deps; it names no rusty-sudocode-cli or
  engine-acp type.
- CliHookProgressReporter (a CLI stderr sink) is NOT moved: the builder now takes
  an injected Option<Box<dyn runtime::HookProgressReporter>> (dependency
  inversion); every engine call site passes None, exactly preserving today's
  behavior (emit_output has been false on every seam RuntimeConfig, so the
  reporter was already never installed). CliHookProgressReporter + emit_output
  stay in the CLI.
- Types the CLI still reads (ModelSwitchReport fields; the BuiltRuntime handle via
  new runtime_mut()/runtime_ref()/mcp_state() accessors) are pub; the field stays
  private. max_tokens_for_model -> engine_core; engine_events::TurnComplete ->
  engine_core::TurnComplete; config resolvers via engine_host::config. Deleted two
  dead builders (build_runtime, build_runtime_plugin_state -- zero callers).

The still-in-CLI ACP delegate (AcpCliAgent/AcpSdkDelegate/run_acp_server) now
uses engine_host:: for the moved symbols; Part B (#119) deletes that layer, after
which the boundary-gate (#121) passes.

Verified: both crates build green + warning-clean, fmt clean, 9 engine-host unit
tests pass, all CLI integration (PTY) tests compile, a live one-shot turn returns
correct output through the moved path. main.rs shrank ~1120 lines. Final staged
sub-step of the engine-host extraction (#120).
