+++ rust/crates/rusty-sudocode-cli/src/main.rs
@@ -126,6 +126,19 @@ use engine_host::session::{
 use engine_host::tool_executor::{
     clear_pending_plan_execution, permission_policy, take_pending_plan_execution, CliToolExecutor,
 };
+// The engine CORE cluster now lives below the seam in `engine-host`
+// (`runtime_build` + `session_engine`, re-exported at the crate root). The
+// renderer names these to build a session (`SessionEngine`), drive its non-turn
+// lifecycle (`SessionLifecycle`), and — on the ACP path only — construct/hold a
+// runtime directly (`BuiltRuntime` / `RuntimeConfig` / `AcpCliSession` + the
+// `build_*` helpers). `RuntimeConfig` is the engine-side config struct; it does
+// not shadow `runtime::RuntimeConfig` (always named fully-qualified).
+use engine_host::{
+    build_engine_runtime, build_plugin_manager, build_runtime_for_cwd,
+    build_runtime_plugin_state_with_loader, plugin_load_outcome_for_cwd, AcpCliSession,
+    BuiltRuntime, ModelSwitchReport, RuntimeConfig, RuntimePluginState, SessionEngine,
+    SessionLifecycle,
+};
 use init::initialize_repo;
 use plugins::{PluginLoadOutcome, PluginManager, PluginRegistry};
 use render::{
@@ -2840,157 +2853,6 @@ struct LiveCli {
     iocraft_output: Option<repl_ui::OutputSender>,
 }
 
-pub(crate) struct RuntimePluginState {
-    pub(crate) feature_config: runtime::RuntimeFeatureConfig,
-    pub(crate) tool_registry: GlobalToolRegistry,
-    pub(crate) plugin_registry: PluginRegistry,
-    pub(crate) plugin_load_outcome: PluginLoadOutcome,
-    pub(crate) mcp_state: Option<Arc<Mutex<RuntimeMcpState>>>,
-}
-
-/// Groups the non-session parameters threaded through the `build_runtime*`
-/// call chain so that adding a new knob only touches one struct instead of
-/// 3-4 function signatures and 10+ call sites.
-#[derive(Clone)]
-struct RuntimeConfig {
-    model: String,
-    system_prompt: SystemPrompt,
-    enable_tools: bool,
-    emit_output: bool,
-    allowed_tools: Option<AllowedToolSet>,
-    permission_mode: PermissionMode,
-    auth_mode: AuthMode,
-    sudocode_config: engine_core::SudoCodeConfig,
-}
-
-struct BuiltRuntime {
-    runtime: Option<ConversationRuntime<EngineApiClient, CliToolExecutor>>,
-    plugin_registry: PluginRegistry,
-    plugin_load_outcome: PluginLoadOutcome,
-    plugins_active: bool,
-    mcp_state: Option<Arc<Mutex<RuntimeMcpState>>>,
-    mcp_active: bool,
-}
-
-impl BuiltRuntime {
-    fn new(
-        runtime: ConversationRuntime<EngineApiClient, CliToolExecutor>,
-        plugin_registry: PluginRegistry,
-        plugin_load_outcome: PluginLoadOutcome,
-        mcp_state: Option<Arc<Mutex<RuntimeMcpState>>>,
-    ) -> Self {
-        Self {
-            runtime: Some(runtime),
-            plugin_registry,
-            plugin_load_outcome,
-            plugins_active: true,
-            mcp_state,
-            mcp_active: true,
-        }
-    }
-
-    fn with_hook_abort_signal(mut self, hook_abort_signal: runtime::HookAbortSignal) -> Self {
-        let runtime = self
-            .runtime
-            .take()
-            .expect("runtime should exist before installing hook abort signal");
-        self.runtime = Some(runtime.with_hook_abort_signal(hook_abort_signal));
-        self
-    }
-
-    fn with_session_known_date(mut self, date: impl Into<String>) -> Self {
-        let runtime = self
-            .runtime
-            .take()
-            .expect("runtime should exist before overriding session known date");
-        self.runtime = Some(runtime.with_session_known_date(date));
-        self
-    }
-
-    /// Set the trace ID for the next request.
-    fn set_trace_id(&mut self, trace_id: impl Into<String>) {
-        if let Some(ref mut runtime) = self.runtime {
-            runtime.set_trace_id(trace_id);
-        }
-    }
-
-    fn plugin_load_outcome(&self) -> &PluginLoadOutcome {
-        &self.plugin_load_outcome
-    }
-
-    fn shutdown_plugins(&mut self) -> Result<(), Box<dyn std::error::Error>> {
-        if self.plugins_active {
-            self.plugin_registry.shutdown()?;
-            self.plugins_active = false;
-        }
-        Ok(())
-    }
-
-    fn shutdown_mcp(&mut self) -> Result<(), Box<dyn std::error::Error>> {
-        if self.mcp_active {
-            if let Some(mcp_state) = &self.mcp_state {
-                mcp_state
-                    .lock()
-                    .unwrap_or_else(std::sync::PoisonError::into_inner)
-                    .shutdown()?;
-            }
-            self.mcp_active = false;
-        }
-        Ok(())
-    }
-
-    /// Returns a reference to the session tracer, if available.
-    fn session_tracer(&self) -> Option<&telemetry::SessionTracer> {
-        self.runtime
-            .as_ref()
-            .expect("runtime should exist while built runtime is alive")
-            .api_client()
-            .session_tracer()
-    }
-}
-
-impl Deref for BuiltRuntime {
-    type Target = ConversationRuntime<EngineApiClient, CliToolExecutor>;
-
-    fn deref(&self) -> &Self::Target {
-        self.runtime
-            .as_ref()
-            .expect("runtime should exist while built runtime is alive")
-    }
-}
-
-impl DerefMut for BuiltRuntime {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        self.runtime
-            .as_mut()
-            .expect("runtime should exist while built runtime is alive")
-    }
-}
-
-impl Drop for BuiltRuntime {
-    fn drop(&mut self) {
-        let _ = self.shutdown_mcp();
-        let _ = self.shutdown_plugins();
-    }
-}
-
-struct AcpCliSession {
-    cwd: PathBuf,
-    handle: SessionHandle,
-    runtime: BuiltRuntime,
-    abort_signal: runtime::HookAbortSignal,
-    /// Session start time for duration tracking.
-    started_at: Instant,
-    /// per-session injected MCP servers (from session/new or session/load),
-    /// reused when the runtime is rebuilt (e.g. model switch) so they
-    /// survive across the session's lifetime.
-    session_mcp_servers: std::collections::BTreeMap<String, runtime::ScopedMcpServerConfig>,
-    /// Caller-supplied system-prompt adjustments (`_meta.sudocode.systemPrompt`
-    /// / `appendSystemPrompt` on session/new or session/load). Kept on the
-    /// session so a runtime rebuild (model switch) re-applies them.
-    prompt_overrides: runtime::SystemPromptOverrides,
-}
-
 /// One live ACP session as held by [`AcpCliAgent`].
 ///
 /// The session proper sits behind its **own** mutex so that turns of
@@ -3117,6 +2979,7 @@ impl AcpCliAgent {
             mcp_servers,
             abort_signal.clone(),
             self.reasoning_effort.clone(),
+            None,
         )
         .map_err(|error| AcpError::internal(format!("failed to build runtime: {error}")))?;
         runtime
@@ -3259,6 +3122,7 @@ impl AcpCliAgent {
             &session_mcp,
             session.abort_signal.clone(),
             self.reasoning_effort.clone(),
+            None,
         )
         .map_err(|e| AcpError::internal(e.to_string()))?;
         Ok(runtime)
@@ -3425,690 +3289,6 @@ fn copy_dir_recursive(from: &Path, to: &Path) -> io::Result<()> {
         }
     }
     Ok(())
-}
-
-/// The CLI's implementation of the seam's [`engine_core::EngineDelegate`] — one
-/// live session's runtime, driven the same way for every renderer (the REPL via
-/// `EngineSession`, ACP over stdio/ws). It owns one session's build →
-/// run-turn (+ auto-compact) → model-switch → persist lifecycle, single-session,
-/// returning the seam's neutral `TurnComplete`. Every renderer shares this one
-/// core — nothing renders here.
-///
-/// The active model is the session's own (`session.model`); this type keeps no
-/// separate copy.
-/// Outcome of a model switch, returned by [`SessionEngine::set_model_impl`] so
-/// each seam consumer formats it its own way: the REPL renders a
-/// `format_model_switch_report` / `format_model_report`; the pump/ACP path takes
-/// `(resolved, available)`. Report DATA only — no formatting crosses the seam.
-pub(crate) struct ModelSwitchReport {
-    /// The model in effect before the switch.
-    previous: String,
-    /// The resolved target model (equal to `previous` when it was a no-op).
-    resolved: String,
-    /// Session message count at switch time (for the report lines).
-    message_count: usize,
-    /// Usage turns at switch time (for the no-op "current model" report).
-    turns: u32,
-    /// `false` when the target equalled the current model — no rebuild happened.
-    changed: bool,
-    /// Config keys ∪ discovery, current model pinned first (for `ModelChanged`).
-    available: Vec<String>,
-}
-
-struct SessionEngine {
-    session: std::sync::Mutex<AcpCliSession>,
-    tokio_runtime: tokio::runtime::Runtime,
-    allowed_tools: Option<AllowedToolSet>,
-    /// Effective permission mode — the engine's SSOT (the renderer no longer
-    /// keeps a copy). Resolved once at build from the CLI override / default;
-    /// `/permissions` mutates it in place and every runtime rebuild reads it.
-    permission_mode: std::sync::Mutex<PermissionMode>,
-    /// Reasoning effort — immutable for now (no `/effort` verb yet).
-    reasoning_effort: Option<String>,
-    /// Auth-mode override — the engine's SSOT. `None` = auto-resolve from the
-    /// model + config; `/auth` pins a concrete mode. Every rebuild reads it.
-    auth_mode: std::sync::Mutex<Option<AuthMode>>,
-}
-
-impl SessionEngine {
-    /// Build a single-session engine for `cwd`. Ports `AcpCliAgent::build_session`.
-    ///
-    /// `system_prompt` is supplied by the caller (the REPL passes its own
-    /// system prompt; the ACP path passes the ACP one) so this one engine core
-    /// serves every renderer without baking in a prompt policy.
-    #[allow(clippy::too_many_arguments)]
-    fn build(
-        cwd: &Path,
-        mcp_servers: &std::collections::BTreeMap<String, runtime::ScopedMcpServerConfig>,
-        prompt_overrides: runtime::SystemPromptOverrides,
-        system_prompt: SystemPrompt,
-        model: String,
-        model_flag_raw: Option<String>,
-        allowed_tools: Option<AllowedToolSet>,
-        permission_mode_override: Option<PermissionMode>,
-        reasoning_effort: Option<String>,
-        auth_mode: Option<AuthMode>,
-    ) -> Result<Self, String> {
-        let cwd = canonical_session_cwd(cwd)?;
-        let _scope = runtime::WorkspaceRootScope::enter(&cwd);
-        let resolved_model = if model_flag_raw.is_some() {
-            model.clone()
-        } else {
-            resolve_repl_model(model.clone())
-        };
-        let permission_mode = permission_mode_override.unwrap_or_else(default_permission_mode);
-        let session_state =
-            new_cli_session_for(&cwd).map_err(|e| format!("failed to create session: {e}"))?;
-        let handle = create_managed_session_handle_for(&cwd, &session_state.session_id)
-            .map_err(|e| format!("failed to create session handle: {e}"))?;
-        let sudocode_config = require_sudocode_config_for_cwd(&cwd)?;
-        let resolved_auth = resolve_auth_mode(&resolved_model, auth_mode, &sudocode_config)
-            .map_err(|e| format!("failed to resolve auth mode: {e}"))?;
-        let abort_signal = runtime::HookAbortSignal::new();
-        let runtime = build_engine_runtime(
-            &cwd,
-            session_state.with_persistence_path(handle.path.clone()),
-            &handle.id,
-            RuntimeConfig {
-                model: resolved_model.clone(),
-                system_prompt,
-                enable_tools: true,
-                emit_output: false,
-                allowed_tools: allowed_tools.clone(),
-                permission_mode,
-                auth_mode: resolved_auth,
-                sudocode_config,
-            },
-            mcp_servers,
-            abort_signal.clone(),
-            reasoning_effort.clone(),
-        )
-        .map_err(|e| format!("failed to build runtime: {e}"))?;
-        runtime
-            .session()
-            .save_to_path(&handle.path)
-            .map_err(|e| format!("failed to persist session: {e}"))?;
-
-        let session = AcpCliSession {
-            cwd,
-            handle,
-            runtime,
-            abort_signal,
-            started_at: Instant::now(),
-            session_mcp_servers: mcp_servers.clone(),
-            prompt_overrides,
-        };
-        Ok(Self {
-            session: std::sync::Mutex::new(session),
-            tokio_runtime: tokio::runtime::Runtime::new()
-                .map_err(|e| format!("failed to create engine tokio runtime: {e}"))?,
-            allowed_tools,
-            permission_mode: std::sync::Mutex::new(permission_mode),
-            reasoning_effort,
-            auth_mode: std::sync::Mutex::new(auth_mode),
-        })
-    }
-
-    fn lock_session(&self) -> std::sync::MutexGuard<'_, AcpCliSession> {
-        self.session
-            .lock()
-            .unwrap_or_else(std::sync::PoisonError::into_inner)
-    }
-
-    /// THE one runtime-rebuild primitive behind every session swap (model / auth
-    /// / permission switch, `/clear`, `/resume`, `/session switch|fork`,
-    /// compaction, plugin reload). Rebuilds the locked session's runtime for
-    /// `new_session` under `handle`, reading the effective config straight from
-    /// the engine's SSOT (`self.permission_mode` / `self.auth_mode` /
-    /// `self.reasoning_effort`) and the model from `new_session.model` — the
-    /// caller sets that field to the intended effective model, or leaves it
-    /// `None` to inherit the current one. Abort signal + prompt overrides come
-    /// from the live session. No model/permission/auth params: the engine, not
-    /// the renderer, owns that config (audit finding B — one SSOT).
-    fn rebuild_locked(
-        &self,
-        session: &mut AcpCliSession,
-        mut new_session: runtime::Session,
-        handle: SessionHandle,
-    ) -> Result<(), String> {
-        let cwd = session.cwd.clone();
-        let _scope = runtime::WorkspaceRootScope::enter(&cwd);
-        if new_session.model.is_none() {
-            new_session.model = session.runtime.session().model.clone();
-        }
-        let model = new_session.model.clone().unwrap_or_default();
-        let permission_mode = self.locked_permission_mode();
-        let auth_override = self.auth_override();
-        let sudocode_config = load_sudocode_config_for_cwd(&cwd);
-        let auth_mode = resolve_model_switch_auth_mode(&model, auth_override, &sudocode_config)
-            .map_err(|e| format!("failed to resolve auth mode: {e}"))?;
-        let system_prompt = build_acp_system_prompt(&cwd, &session.prompt_overrides)?;
-        let runtime = build_engine_runtime(
-            &cwd,
-            new_session,
-            &handle.id,
-            RuntimeConfig {
-                model,
-                system_prompt,
-                enable_tools: true,
-                emit_output: false,
-                allowed_tools: self.allowed_tools.clone(),
-                permission_mode,
-                auth_mode,
-                sudocode_config,
-            },
-            &session.session_mcp_servers,
-            session.abort_signal.clone(),
-            self.reasoning_effort.clone(),
-        )
-        .map_err(|e| e.to_string())?;
-        session.runtime = runtime;
-        session.handle = handle;
-        Ok(())
-    }
-
-    /// Effective permission mode (engine SSOT). Private helper behind the
-    /// `SessionLifecycle::current_permission_mode` read + every rebuild.
-    fn locked_permission_mode(&self) -> PermissionMode {
-        *self
-            .permission_mode
-            .lock()
-            .unwrap_or_else(std::sync::PoisonError::into_inner)
-    }
-
-    /// The raw auth-mode override (`None` = auto-resolve). Internal — callers
-    /// wanting the resolved mode use [`Self::resolved_auth_mode`].
-    fn auth_override(&self) -> Option<AuthMode> {
-        *self
-            .auth_mode
-            .lock()
-            .unwrap_or_else(std::sync::PoisonError::into_inner)
-    }
-
-    /// The model in effect (engine SSOT = the session's own model).
-    fn session_model(&self) -> String {
-        self.lock_session()
-            .runtime
-            .session()
-            .model
-            .clone()
-            .unwrap_or_default()
-    }
-
-    /// The resolved auth mode in effect: the pinned override, else the mode the
-    /// current model + config auto-resolve to (what the runtime actually uses).
-    fn resolved_auth_mode(&self) -> AuthMode {
-        let model = self.session_model();
-        let config = load_sudocode_config_for_current_dir();
-        resolve_auth_mode(&model, self.auth_override(), &config).unwrap_or(AuthMode::ApiKey)
-    }
-
-    /// Config keys ∪ discovery ids, `current` pinned first — the model list the
-    /// seam's `ModelChanged` carries and `/model` (no arg) shows.
-    fn available_models(&self, current: &str) -> Vec<String> {
-        let config = load_sudocode_config_for_current_dir();
-        let config_keys: Vec<String> = config.models.keys().cloned().collect();
-        let mut available = runtime::model_capabilities::merge_discovery_ids(&config_keys);
-        if !available.iter().any(|m| m.eq_ignore_ascii_case(current)) {
-            available.insert(0, current.to_string());
-        }
-        available
-    }
-
-    /// The one model-switch implementation, shared by the turn seam
-    /// ([`engine_core::EngineDelegate::set_model`]) and the REPL lifecycle
-    /// ([`SessionLifecycle::set_model`]) — audit finding A (was 4 copies).
-    /// Resolves the alias, and when it differs from the current model rebuilds
-    /// the runtime via [`Self::rebuild_locked`] (keeping `session.model` in sync
-    /// so auto-compaction reads the right context window). Returns report DATA;
-    /// neither caller formats here.
-    fn set_model_impl(&self, new_model: &str) -> Result<ModelSwitchReport, String> {
-        let resolved = resolve_model_alias_with_config(new_model);
-        let (previous, message_count, turns, changed) = {
-            let mut session = self.lock_session();
-            let _scope = runtime::WorkspaceRootScope::enter(&session.cwd);
-            let previous = session.runtime.session().model.clone().unwrap_or_default();
-            let message_count = session.runtime.session().messages.len();
-            let turns = session.runtime.usage().turns();
-            if resolved == previous {
-                (previous, message_count, turns, false)
-            } else {
-                let mut new_session = session.runtime.session().clone();
-                new_session.model = Some(resolved.clone());
-                let handle = session.handle.clone();
-                self.rebuild_locked(&mut session, new_session, handle)?;
-                (previous, message_count, turns, true)
-            }
-        };
-        let available = self.available_models(&resolved);
-        Ok(ModelSwitchReport {
-            previous,
-            resolved,
-            message_count,
-            turns,
-            changed,
-            available,
-        })
-    }
-
-    /// The one permission-mode implementation, shared by the seam and the REPL:
-    /// update the engine SSOT, then flip the live policy's active mode in place
-    /// (the lightweight mechanism the ACP path already trusts — no full rebuild).
-    fn set_permission_mode_impl(&self, mode: PermissionMode) {
-        *self
-            .permission_mode
-            .lock()
-            .unwrap_or_else(std::sync::PoisonError::into_inner) = mode;
-        let mut session = self.lock_session();
-        if let Some(rt) = session.runtime.runtime.as_mut() {
-            rt.permission_policy_mut().set_active_mode(mode);
-        }
-    }
-}
-
-impl engine_core::EngineDelegate for SessionEngine {
-    fn run_turn(
-        &self,
-        blocks: Vec<runtime::ContentBlock>,
-        observer: &mut dyn runtime::RuntimeObserver,
-        prompter: &mut dyn runtime::PermissionPrompter,
-    ) -> Result<engine_events::TurnComplete, String> {
-        let mut session = self.lock_session();
-        session.abort_signal.reset();
-        let _scope = runtime::WorkspaceRootScope::enter(&session.cwd);
-
-        // Pre-send auto-compaction, budgeted the way the API preflight is
-        // (context window minus max output, the fixed per-request overhead, and
-        // the autocompact buffer) so a too-large request never reaches the
-        // provider — the #545 wedge fix, on the engine turn path. The session's
-        // own model is the SSOT for the context-window lookup (build + set_model
-        // keep it current). Compacts through the LLM path and rewrites the
-        // persisted transcript; the runtime also compacts-and-resends once
-        // reactively inside run_turn if the provider still rejects.
-        let model = session.runtime.session().model.clone().unwrap_or_default();
-        let context_limit = runtime::model_capabilities::context_window_or_default(&model) as usize;
-        let max_output_tokens = max_tokens_for_model(&model) as usize;
-        let overhead_tokens = session
-            .runtime
-            .api_client()
-            .fixed_request_overhead_tokens(session.runtime.system_prompt());
-        let buffer_tokens = runtime::autocompact_buffer_tokens(&model) as usize;
-        let history_budget =
-            context_limit.saturating_sub(max_output_tokens + overhead_tokens + buffer_tokens);
-        let prompt_tokens: usize = blocks.iter().map(estimate_block_tokens).sum();
-        let estimated_tokens = estimate_session_tokens(session.runtime.session());
-        let mut pre_send_compaction = None;
-        if estimated_tokens + prompt_tokens > history_budget {
-            if let Some(tracer) = session.runtime.session_tracer() {
-                tracer.record("auto_compact_check", {
-                    let mut attrs = Map::new();
-                    attrs.insert(
-                        "estimated_tokens".to_string(),
-                        Value::Number(estimated_tokens.into()),
-                    );
-                    attrs.insert(
-                        "prompt_tokens".to_string(),
-                        Value::Number(prompt_tokens.into()),
-                    );
-                    attrs.insert(
-                        "history_budget".to_string(),
-                        Value::Number(history_budget.into()),
-                    );
-                    attrs.insert(
-                        "context_limit".to_string(),
-                        Value::Number(context_limit.into()),
-                    );
-                    attrs
-                });
-            }
-            pre_send_compaction = self
-                .tokio_runtime
-                .block_on(session.runtime.compact_in_place(CompactionConfig {
-                    max_estimated_tokens: 0, // force compaction
-                    ..CompactionConfig::default()
-                }));
-            // Re-estimate against the hard limit the preflight enforces. Still
-            // over → classified error instead of a request that will be rejected.
-            let new_estimated_tokens = estimate_session_tokens(session.runtime.session());
-            if new_estimated_tokens + prompt_tokens + overhead_tokens + max_output_tokens
-                > context_limit
-            {
-                return Err(context_overflow_user_message(
-                    session.runtime.session(),
-                    new_estimated_tokens + prompt_tokens + overhead_tokens + max_output_tokens,
-                    context_limit,
-                ));
-            }
-        }
-
-        let turn_summary = self
-            .tokio_runtime
-            .block_on(
-                session
-                    .runtime
-                    .run_turn_with_blocks(blocks, Some(prompter), Some(observer)),
-            )
-            .map_err(|e| e.to_string())?;
-
-        let path = session.handle.path.clone();
-        session
-            .runtime
-            .session()
-            .save_to_path(&path)
-            .map_err(|e| format!("failed to persist session: {e}"))?;
-
-        Ok(engine_events::TurnComplete {
-            iterations: turn_summary.iterations,
-            turn_usage: turn_summary.turn_usage,
-            session_usage: turn_summary.session_usage,
-            cancelled: turn_summary.cancelled,
-            response_model: turn_summary.response_model,
-            // Prefer the pre-send compaction event; else the runtime's in-turn one.
-            auto_compaction: pre_send_compaction.or(turn_summary.auto_compaction),
-        })
-    }
-
-    fn set_question_prompter(&self, prompter: Box<dyn runtime::QuestionPrompter>) {
-        let mut session = self.lock_session();
-        if let Some(rt) = session.runtime.runtime.as_mut() {
-            rt.tool_executor_mut().set_question_prompter(prompter);
-        }
-    }
-
-    fn abort_signal(&self) -> runtime::HookAbortSignal {
-        self.lock_session().abort_signal.clone()
-    }
-
-    fn set_model(&self, new_model: &str) -> Result<(String, Vec<String>), String> {
-        let report = self.set_model_impl(new_model)?;
-        Ok((report.resolved, report.available))
-    }
-
-    fn set_permission_mode(&self, mode: PermissionMode) -> Result<(), String> {
-        self.set_permission_mode_impl(mode);
-        Ok(())
-    }
-
-    fn handle_slash_command(&self, line: &str) -> Result<String, String> {
-        // `/model <name>` switches the model; other slash commands are handled by
-        // the renderer locally (the REPL intercepts them before they reach the
-        // engine). Kept minimal here; the seam only needs the model verb.
-        let trimmed = line.trim();
-        if let Some(rest) = trimmed.strip_prefix("/model") {
-            let arg = rest.trim();
-            if arg.is_empty() {
-                let current = self
-                    .lock_session()
-                    .runtime
-                    .session()
-                    .model
-                    .clone()
-                    .unwrap_or_default();
-                return Ok(format!("current model: {current}"));
-            }
-            let resolved = self.set_model_impl(arg)?.resolved;
-            return Ok(format!("switched model to {resolved}"));
-        }
-        Ok(String::new())
-    }
-
-    fn close(&self) {
-        let session = self.lock_session();
-        let path = session.handle.path.clone();
-        let _ = session.runtime.session().save_to_path(&path);
-    }
-}
-
-/// The non-turn session-lifecycle contract the composition root (`LiveCli`) uses
-/// to manage a live engine session **without being able to drive turns** — turns
-/// go only through `EngineHandle`. Split from [`engine_core::EngineDelegate`]
-/// (turn ops) so a renderer physically cannot bypass the seam: holding an
-/// `Arc<dyn SessionLifecycle>` gives no access to `run_turn`. SRP: turn-driving
-/// and session-management are two orthogonal responsibilities.
-///
-/// Methods are dyn-safe (no generics) and return owned snapshots — these are
-/// rare, non-hot-path management ops (export / status / slash / undo), so the
-/// clones cost nothing on the critical path.
-pub(crate) trait SessionLifecycle: Send + Sync + 'static {
-    /// A clone of the current session (for export / status / read inspection).
-    fn session_snapshot(&self) -> runtime::Session;
-    /// The active session's handle (id + persistence path). Owned clone so a
-    /// slash-command handler can name the session without touching the runtime
-    /// mid-turn.
-    fn session_handle(&self) -> SessionHandle;
-    /// Persist the session to its backing path.
-    fn persist(&self) -> Result<(), String>;
-    /// Mutate the session in place (undo, fork prep, …).
-    fn with_session_mut(&self, f: &mut dyn FnMut(&mut runtime::Session));
-    /// Snapshot of the runtime's cumulative/turn usage tracker (for
-    /// `/status`, `/cost`, `/stats`, resume/model reports). Cloned, not
-    /// borrowed, so it never pins the session lock.
-    fn usage_snapshot(&self) -> runtime::UsageTracker;
-    /// Estimated token footprint of the current session (for `/status`).
-    fn estimated_tokens(&self) -> usize;
-    /// A clone of the session tracer, if telemetry is active. Returned owned
-    /// (the tracer is `Arc`-backed and cheap to clone) so callers record events
-    /// without holding the session lock.
-    fn session_tracer(&self) -> Option<telemetry::SessionTracer>;
-    /// A snapshot of the plugin load outcome (for `/skills` resolution).
-    fn plugin_load_outcome(&self) -> PluginLoadOutcome;
-    /// Run an `/mcp reconnect|enable|disable <server>` action against the live
-    /// MCP state. `None` when no MCP servers are running in this session; else
-    /// the action's `Ok(message)` / `Err(message)`.
-    fn mcp_command(&self, action: &str, server: &str) -> Option<Result<String, String>>;
-
-    // --- config reads (engine SSOT) ------------------------------------------
-    /// The model in effect.
-    fn current_model(&self) -> String;
-    /// The effective permission mode.
-    fn current_permission_mode(&self) -> PermissionMode;
-    /// The resolved auth mode in effect.
-    fn current_auth_mode(&self) -> AuthMode;
-
-    // --- semantic session ops (engine owns the rebuild; renderer only formats)-
-    /// Switch the model. Returns report DATA (`previous` / `resolved` /
-    /// `changed` / counts / `available`); the renderer formats it. Shares the
-    /// one impl with [`engine_core::EngineDelegate::set_model`].
-    fn set_model(&self, new_model: &str) -> Result<ModelSwitchReport, String>;
-    /// Pin the auth mode (engine SSOT) and rebuild so it takes effect.
-    fn set_auth(&self, mode: AuthMode) -> Result<(), String>;
-    /// Switch the active permission mode (engine SSOT) in place.
-    fn set_permission_mode(&self, mode: PermissionMode) -> Result<(), String>;
-    /// Start a fresh session (`/clear`), preserving the current model. Returns
-    /// the new session handle for the renderer to adopt + report.
-    fn reset_session(&self) -> Result<SessionHandle, String>;
-    /// Resume a session by reference (`/resume`, `/session switch`): the engine
-    /// loads it and swaps it in, keeping the current effective model. Returns
-    /// the new handle + its message count.
-    fn resume_session(&self, reference: &str) -> Result<(SessionHandle, usize), String>;
-    /// Fork the current session (`/session fork`). Returns the new handle, its
-    /// message count, and the branch name (if any).
-    fn fork_session(
-        &self,
-        branch: Option<String>,
-    ) -> Result<(SessionHandle, usize, Option<String>), String>;
-    /// Rebuild the runtime to pick up reloaded plugin / feature state, then
-    /// persist (`/plugins` reload).
-    fn reload_features(&self) -> Result<(), String>;
-    /// Run LLM-based history compaction and swap the compacted session in.
-    /// Returns `(removed, kept, skipped)` for the renderer's report.
-    fn run_compaction(
-        &self,
-    ) -> Result<(usize, usize, bool, runtime::CompactionSummarySource), String>;
-}
-
-impl SessionLifecycle for SessionEngine {
-    fn session_snapshot(&self) -> runtime::Session {
-        self.lock_session().runtime.session().clone()
-    }
-
-    fn session_handle(&self) -> SessionHandle {
-        self.lock_session().handle.clone()
-    }
-
-    fn persist(&self) -> Result<(), String> {
-        let session = self.lock_session();
-        let path = session.handle.path.clone();
-        session
-            .runtime
-            .session()
-            .save_to_path(&path)
-            .map_err(|e| e.to_string())
-    }
-
-    fn with_session_mut(&self, f: &mut dyn FnMut(&mut runtime::Session)) {
-        let mut session = self.lock_session();
-        f(session.runtime.session_mut());
-    }
-
-    fn usage_snapshot(&self) -> runtime::UsageTracker {
-        self.lock_session().runtime.usage().clone()
-    }
-
-    fn estimated_tokens(&self) -> usize {
-        self.lock_session().runtime.estimated_tokens()
-    }
-
-    fn session_tracer(&self) -> Option<telemetry::SessionTracer> {
-        self.lock_session().runtime.session_tracer().cloned()
-    }
-
-    fn plugin_load_outcome(&self) -> PluginLoadOutcome {
-        self.lock_session().runtime.plugin_load_outcome().clone()
-    }
-
-    fn mcp_command(&self, action: &str, server: &str) -> Option<Result<String, String>> {
-        let session = self.lock_session();
-        let mcp_state = session.runtime.mcp_state.as_ref()?;
-        let mut mcp = mcp_state
-            .lock()
-            .unwrap_or_else(std::sync::PoisonError::into_inner);
-        let result = match action {
-            "reconnect" => mcp.reconnect_server(server),
-            "enable" => mcp.enable_server(server),
-            "disable" => mcp.disable_server(server),
-            other => return Some(Err(format!("unknown /mcp action: {other}"))),
-        };
-        Some(result.map_err(|e| e.to_string()))
-    }
-
-    fn current_model(&self) -> String {
-        self.session_model()
-    }
-
-    fn current_permission_mode(&self) -> PermissionMode {
-        self.locked_permission_mode()
-    }
-
-    fn current_auth_mode(&self) -> AuthMode {
-        self.resolved_auth_mode()
-    }
-
-    fn set_model(&self, new_model: &str) -> Result<ModelSwitchReport, String> {
-        self.set_model_impl(new_model)
-    }
-
-    fn set_auth(&self, mode: AuthMode) -> Result<(), String> {
-        *self
-            .auth_mode
-            .lock()
-            .unwrap_or_else(std::sync::PoisonError::into_inner) = Some(mode);
-        let mut session = self.lock_session();
-        let new_session = session.runtime.session().clone();
-        let handle = session.handle.clone();
-        self.rebuild_locked(&mut session, new_session, handle)
-    }
-
-    fn set_permission_mode(&self, mode: PermissionMode) -> Result<(), String> {
-        self.set_permission_mode_impl(mode);
-        Ok(())
-    }
-
-    fn reset_session(&self) -> Result<SessionHandle, String> {
-        let mut session = self.lock_session();
-        let _scope = runtime::WorkspaceRootScope::enter(&session.cwd);
-        let current_model = session.runtime.session().model.clone();
-        let session_state = new_cli_session_for(&session.cwd).map_err(|e| e.to_string())?;
-        let handle = create_managed_session_handle_for(&session.cwd, &session_state.session_id)
-            .map_err(|e| e.to_string())?;
-        let mut fresh = session_state.with_persistence_path(handle.path.clone());
-        fresh.model = current_model;
-        self.rebuild_locked(&mut session, fresh, handle.clone())?;
-        Ok(handle)
-    }
-
-    fn resume_session(&self, reference: &str) -> Result<(SessionHandle, usize), String> {
-        let mut session = self.lock_session();
-        let _scope = runtime::WorkspaceRootScope::enter(&session.cwd);
-        let (handle, mut loaded) = load_session_reference(reference).map_err(|e| e.to_string())?;
-        let message_count = loaded.messages.len();
-        // Keep the current effective model (REPL parity: the runtime model is
-        // config-driven, not adopted from the resumed session).
-        loaded.model = session.runtime.session().model.clone();
-        self.rebuild_locked(&mut session, loaded, handle.clone())?;
-        Ok((handle, message_count))
-    }
-
-    fn fork_session(
-        &self,
-        branch: Option<String>,
-    ) -> Result<(SessionHandle, usize, Option<String>), String> {
-        let mut session = self.lock_session();
-        let _scope = runtime::WorkspaceRootScope::enter(&session.cwd);
-        let forked = session.runtime.fork_session(branch);
-        let handle = create_managed_session_handle_for(&session.cwd, &forked.session_id)
-            .map_err(|e| e.to_string())?;
-        let branch_name = forked
-            .fork
-            .as_ref()
-            .and_then(|fork| fork.branch_name.clone());
-        let forked = forked.with_persistence_path(handle.path.clone());
-        let message_count = forked.messages.len();
-        forked
-            .save_to_path(&handle.path)
-            .map_err(|e| e.to_string())?;
-        self.rebuild_locked(&mut session, forked, handle.clone())?;
-        Ok((handle, message_count, branch_name))
-    }
-
-    fn reload_features(&self) -> Result<(), String> {
-        let mut session = self.lock_session();
-        let new_session = session.runtime.session().clone();
-        let handle = session.handle.clone();
-        self.rebuild_locked(&mut session, new_session, handle)?;
-        let path = session.handle.path.clone();
-        session
-            .runtime
-            .session()
-            .save_to_path(&path)
-            .map_err(|e| e.to_string())
-    }
-
-    fn run_compaction(
-        &self,
-    ) -> Result<(usize, usize, bool, runtime::CompactionSummarySource), String> {
-        let mut session = self.lock_session();
-        let cwd = session.cwd.clone();
-        let _scope = runtime::WorkspaceRootScope::enter(&cwd);
-        let result = self
-            .tokio_runtime
-            .block_on(session.runtime.compact(CompactionConfig::default(), None));
-        let removed = result.removed_message_count;
-        let kept = result.compacted_session.messages.len();
-        let skipped = removed == 0;
-        // Surface the summary provenance to the renderer's `/compact` report
-        // (LLM vs heuristic) — main's compaction hardening added this column.
-        let summary_source = result.summary_source;
-        let handle = session.handle.clone();
-        self.rebuild_locked(&mut session, result.compacted_session, handle)?;
-        let path = session.handle.path.clone();
-        session
-            .runtime
-            .session()
-            .save_to_path(&path)
-            .map_err(|e| e.to_string())?;
-        Ok((removed, kept, skipped, summary_source))
-    }
 }
 
 fn run_acp_server(
@@ -4583,7 +3763,7 @@ impl engine_acp::acp_sdk_server::SdkAcpDelegate for AcpSdkDelegate {
     ) -> Result<(), engine_acp::AcpError> {
         let locked = self.lock_session(session_id)?;
         let mut session = locked.get();
-        if let Some(rt) = session.runtime.runtime.as_mut() {
+        if let Some(rt) = session.runtime.runtime_mut() {
             rt.permission_policy_mut().set_active_mode(mode);
         }
         Ok(())
@@ -4837,12 +4017,13 @@ impl AcpSdkDelegate {
                 sudocode_config,
             },
             &mcp_servers,
+            None,
         )
         .map_err(|e| engine_acp::AcpError::internal(format!("failed to build runtime: {e}")))?;
 
         let abort_signal = runtime::HookAbortSignal::new();
         runtime = runtime.with_hook_abort_signal(abort_signal.clone());
-        if let Some(rt) = runtime.runtime.as_mut() {
+        if let Some(rt) = runtime.runtime_mut() {
             rt.api_client_mut()
                 .set_reasoning_effort(self.inner.reasoning_effort.clone());
             let thinking = ConfigLoader::default_for(&cwd)
@@ -7086,97 +6267,6 @@ fn build_system_prompt() -> Result<SystemPrompt, Box<dyn std::error::Error>> {
     build_system_prompt_for(&env::current_dir()?)
 }
 
-fn build_runtime_plugin_state() -> Result<RuntimePluginState, Box<dyn std::error::Error>> {
-    let cwd = env::current_dir()?;
-    let loader = ConfigLoader::default_for(&cwd);
-    let runtime_config = loader.load()?;
-    let session_mcp: std::collections::BTreeMap<String, runtime::ScopedMcpServerConfig> =
-        std::collections::BTreeMap::new();
-    build_runtime_plugin_state_with_loader(&cwd, &loader, &runtime_config, &session_mcp)
-}
-
-fn plugin_load_outcome_for_cwd(
-    cwd: &Path,
-) -> Result<PluginLoadOutcome, Box<dyn std::error::Error>> {
-    let loader = ConfigLoader::default_for(cwd);
-    let runtime_config = loader.load()?;
-    let plugin_manager = build_plugin_manager(cwd, &loader, &runtime_config);
-    Ok(plugin_manager.plugin_registry_report()?.load_outcome())
-}
-
-pub(crate) fn build_runtime_plugin_state_with_loader(
-    cwd: &Path,
-    loader: &ConfigLoader,
-    runtime_config: &runtime::RuntimeConfig,
-    session_mcp: &std::collections::BTreeMap<String, runtime::ScopedMcpServerConfig>,
-) -> Result<RuntimePluginState, Box<dyn std::error::Error>> {
-    // Surface the settings `experimental` section to the process-global
-    // experiments registry BEFORE anything consults a flag (the MCP gate
-    // below is one consumer). First call wins; ACP session rebuilds
-    // against the same config are no-ops.
-    runtime::experiments::init_config_flags(runtime_config.experiments().clone());
-    let plugin_manager = build_plugin_manager(cwd, loader, runtime_config);
-    let plugin_registry_report = plugin_manager.plugin_registry_report()?;
-    let plugin_load_outcome = plugin_registry_report.load_outcome();
-    let plugin_registry = plugin_registry_report.into_registry()?;
-    let plugin_hook_config =
-        runtime_hook_config_from_plugin_hooks(plugin_registry.projected_hooks()?);
-    let feature_config = runtime_config
-        .feature_config()
-        .clone()
-        .with_hooks(runtime_config.hooks().merged(&plugin_hook_config));
-    let tool_registry = GlobalToolRegistry::with_plugin_tools(plugin_registry.aggregated_tools()?)?;
-    let (mcp_state, runtime_tools) =
-        build_runtime_mcp_state(runtime_config, &plugin_load_outcome, session_mcp)?;
-    let tool_registry = match tool_registry.with_runtime_tools(runtime_tools) {
-        Ok(tool_registry) => tool_registry,
-        Err(error) => {
-            shutdown_mcp_state_best_effort(&mcp_state);
-            return Err(Box::new(std::io::Error::other(error)));
-        }
-    };
-    Ok(RuntimePluginState {
-        feature_config,
-        tool_registry,
-        plugin_registry,
-        plugin_load_outcome,
-        mcp_state,
-    })
-}
-
-fn build_plugin_manager(
-    cwd: &Path,
-    loader: &ConfigLoader,
-    runtime_config: &runtime::RuntimeConfig,
-) -> PluginManager {
-    let plugin_config = runtime_config
-        .plugins()
-        .to_plugin_manager_config(cwd, loader.config_home());
-    PluginManager::new(plugin_config)
-}
-
-fn runtime_hook_config_from_plugin_hooks(
-    hooks: plugins::ProjectedPluginHooks,
-) -> runtime::RuntimeHookConfig {
-    runtime::RuntimeHookConfig::new_with_sources(
-        hooks
-            .pre_tool_use
-            .into_iter()
-            .map(|entry| (entry.command, entry.plugin_id))
-            .collect(),
-        hooks
-            .post_tool_use
-            .into_iter()
-            .map(|entry| (entry.command, entry.plugin_id))
-            .collect(),
-        hooks
-            .post_tool_use_failure
-            .into_iter()
-            .map(|entry| (entry.command, entry.plugin_id))
-            .collect(),
-    )
-}
-
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub(crate) struct InternalPromptProgressState {
     pub(crate) command_label: &'static str,
@@ -7398,198 +6488,6 @@ impl Drop for InternalPromptProgressRun {
     fn drop(&mut self) {
         self.stop_heartbeat();
     }
-}
-
-fn build_runtime(
-    session: Session,
-    session_id: &str,
-    config: RuntimeConfig,
-) -> Result<BuiltRuntime, Box<dyn std::error::Error>> {
-    let cwd = env::current_dir()?;
-    let session_mcp: std::collections::BTreeMap<String, runtime::ScopedMcpServerConfig> =
-        std::collections::BTreeMap::new();
-    build_runtime_for_cwd(&cwd, session, session_id, config, &session_mcp)
-}
-
-/// The one place the post-build tail every engine-side runtime (re)build shares
-/// lives: build the runtime for `session` under `handle_id`, install the hook
-/// abort signal, then apply reasoning-effort + `thinking` config. Callers own
-/// the `RuntimeConfig` assembly (their config sources differ) and the
-/// surrounding save/tracer logic; this collapses the identical ~15-line tail
-/// that had been copied across `SessionEngine::build` / `rebuild_locked` /
-/// `set_model_impl` and `AcpCliAgent::build_session` / `handle_acp_model_switch`.
-/// The workspace-root scope must already be active for `cwd`.
-fn build_engine_runtime(
-    cwd: &Path,
-    session: Session,
-    handle_id: &str,
-    config: RuntimeConfig,
-    session_mcp: &std::collections::BTreeMap<String, runtime::ScopedMcpServerConfig>,
-    abort_signal: runtime::HookAbortSignal,
-    reasoning_effort: Option<String>,
-) -> Result<BuiltRuntime, Box<dyn std::error::Error>> {
-    let mut runtime = build_runtime_for_cwd(cwd, session, handle_id, config, session_mcp)?;
-    runtime = runtime.with_hook_abort_signal(abort_signal);
-    if let Some(rt) = runtime.runtime.as_mut() {
-        rt.api_client_mut().set_reasoning_effort(reasoning_effort);
-        let thinking = ConfigLoader::default_for(cwd)
-            .load()
-            .map_or(true, |cfg| cfg.thinking());
-        rt.api_client_mut().set_thinking_enabled(thinking);
-    }
-    Ok(runtime)
-}
-
-fn build_runtime_for_cwd(
-    cwd: &Path,
-    session: Session,
-    session_id: &str,
-    config: RuntimeConfig,
-    session_mcp: &std::collections::BTreeMap<String, runtime::ScopedMcpServerConfig>,
-) -> Result<BuiltRuntime, Box<dyn std::error::Error>> {
-    let loader = ConfigLoader::default_for(cwd);
-    let file_config = loader.load()?;
-    let runtime_plugin_state =
-        build_runtime_plugin_state_with_loader(cwd, &loader, &file_config, session_mcp)?;
-    build_runtime_with_plugin_state(
-        cwd,
-        session,
-        session_id,
-        config,
-        runtime_plugin_state,
-        session_mcp,
-    )
-}
-
-fn build_runtime_with_plugin_state(
-    cwd: &Path,
-    mut session: Session,
-    session_id: &str,
-    mut config: RuntimeConfig,
-    runtime_plugin_state: RuntimePluginState,
-    session_mcp: &std::collections::BTreeMap<String, runtime::ScopedMcpServerConfig>,
-) -> Result<BuiltRuntime, Box<dyn std::error::Error>> {
-    // Persist the model in session metadata so resumed sessions can report it.
-    if session.model.is_none() {
-        session.model = Some(config.model.clone());
-    }
-    let RuntimePluginState {
-        feature_config,
-        tool_registry,
-        plugin_registry,
-        plugin_load_outcome,
-        mcp_state,
-    } = runtime_plugin_state;
-    // Resolve the standalone nexus-A2A session once (fail loud on a partial
-    // config or a dial failure). `None` when A2A is off — the fast path that
-    // leaves scode behaviour unchanged. Held as `Option<&'static Session>`
-    // (Copy) and reused below to advertise, prompt, and wire the send half.
-    let a2a = match engine_host::nexus_a2a::session() {
-        Ok(a2a) => a2a,
-        Err(error) => {
-            shutdown_mcp_state_best_effort(&mcp_state);
-            return Err(Box::new(std::io::Error::other(error)));
-        }
-    };
-    // per-session injected MCP tools bypass the global --allowed-tools gate:
-    // they are explicitly requested for this session and their names are only
-    // known at runtime, so add their qualified names to the allow-list when
-    // one is active. The prefix uses `runtime::mcp_tool_prefix`, which
-    // normalizes server names the same way the tool index does (e.g.
-    // `github.com` -> `mcp__github_com__`), so non-alphanumeric server names
-    // are matched correctly.
-    if let Some(allowed) = config.allowed_tools.as_mut() {
-        if let Some(mcp_state) = &mcp_state {
-            let tools = mcp_state
-                .lock()
-                .unwrap_or_else(std::sync::PoisonError::into_inner)
-                .tools_with_server();
-            allowed.extend(session_mcp_tool_names(tools, session_mcp));
-        }
-    }
-    // nexus A2A: when configured, keep the peer-reply tool available even under
-    // an explicit --allowedTools restriction (absent a restriction it is
-    // already advertised). Its handler is the CliToolExecutor intercept wired
-    // below; the co-host advertises the same tool the same way.
-    if a2a.is_some() {
-        if let Some(allowed) = config.allowed_tools.as_mut() {
-            allowed.extend(["send_message".to_string()]);
-        }
-    }
-    let policy =
-        match permission_policy(config.permission_mode, &feature_config, &tool_registry, cwd) {
-            Ok(policy) => policy,
-            Err(error) => {
-                shutdown_mcp_state_best_effort(&mcp_state);
-                return Err(Box::new(std::io::Error::other(error)));
-            }
-        };
-    let mut system_prompt = config.system_prompt.clone();
-    // Skills are listed so the model can name and load one without the user
-    // having to know it exists. Plugin-provided skill roots are included via
-    // `plugin_load_outcome`, so a plugin can inject skills that the prompt then
-    // advertises. This runs for the REPL, `--print`, and ACP sessions alike:
-    // they all land in this function via `build_runtime_for_cwd`.
-    if let Some(section) = render_skills_prompt_section(cwd, Some(&plugin_load_outcome)) {
-        system_prompt.dynamic_sections.push(section);
-    }
-    // nexus A2A: teach the model its A2A identity + how to reach peers, so the
-    // standalone loop knows it can `send_message` to a named peer.
-    if let Some(session) = a2a {
-        system_prompt
-            .dynamic_sections
-            .push(session.peer_system_prompt());
-    }
-    let emit_output = config.emit_output;
-    let client = match EngineApiClient::new(
-        session_id,
-        &config.sudocode_config,
-        &config.model,
-        config.auth_mode,
-        tool_registry.clone(),
-        config.enable_tools,
-        config.allowed_tools.clone(),
-    ) {
-        Ok(client) => client,
-        Err(error) => {
-            shutdown_mcp_state_best_effort(&mcp_state);
-            return Err(error);
-        }
-    };
-    let mut runtime = ConversationRuntime::new_with_features(
-        session,
-        client,
-        CliToolExecutor::new(
-            config.allowed_tools,
-            tool_registry.clone(),
-            mcp_state.clone(),
-        ),
-        policy,
-        system_prompt,
-        &feature_config,
-    )
-    .with_session_known_date(runtime::today_local());
-    // nexus A2A: give the CLI executor the send half so `send_message` routes
-    // to the peer's replicated DT_STREAM inbox (the shared handler the co-host
-    // uses). Set only when configured; absent it the tool is never advertised.
-    if let Some(session) = a2a {
-        runtime
-            .tool_executor_mut()
-            .set_mailbox_sender(session.sender());
-    }
-    if emit_output {
-        runtime = runtime.with_hook_progress_reporter(Box::new(CliHookProgressReporter));
-    }
-    if let Err(error) = plugin_registry.initialize() {
-        shutdown_mcp_state_best_effort(&mcp_state);
-        return Err(Box::new(error));
-    }
-    Ok(BuiltRuntime::new(
-        runtime,
-        plugin_registry,
-        plugin_load_outcome,
-        mcp_state,
-    ))
 }
 
 struct CliHookProgressReporter;
